fixed: ExplosionComponent.java:[77,37] type parameters of <T>T cannot be determined; no unique maximal instance exists for type variable T with upper bounds float,java.lang.Object
git-svn-id: file:///d/Documents/workspace/gemserksvn/var/svn/games@104 1dcdb8b7-3cac-4752-9c2b-54fd448f9028

+++ towerofdefense/trunk/game/src/main/java/com/gemserk/games/towerofdefense/components/ExplosionComponent.java
@@ -74,7 +74,7 @@ public class ExplosionComponent extends FieldsReflectionComponent {
 		ArrayList<Entity> targets = Properties.getValue(message, "targets");
 
 		for (Entity target : targets) {
-			float damage = Properties.getValue(source, "damage");
+			Float damage = (Float) Properties.getValue(source, "damage");
 			int count = (int) Math.ceil(damage);
 
 			Vector2f position = Properties.getValue(target, "position");
